Fixed the errors in passing by adding the fixturees for plotting into deltakit-explore/tests/plotting/
